refactor: split search page in a new module
