+++ app/build.gradle
@@ -39,12 +39,11 @@ android {
         viewBinding true
     }
 
-    dynamicFeatures = [':favorite', ':onboarding']
+    dynamicFeatures = [':book', ':favorite', ':onboarding', ':search_book']
 }
 
 dependencies {
     implementation project(":core")
-    implementation project(":book")
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
     implementation "com.google.android.play:core:1.10.3"
