+++ app/src/main/java/id/myone/capstone_books_store/ApplicationBase.kt
@@ -6,8 +6,6 @@
 package id.myone.capstone_books_store
 
 import com.google.android.play.core.splitcompat.SplitCompatApplication
-import id.myone.book.di.bookUseCaseModel
-import id.myone.book.di.bookViewModelModule
 import id.myone.capstone_books_store.di.appViewModelModule
 import id.myone.core.di.*
 import org.koin.android.ext.koin.androidContext
@@ -28,9 +26,7 @@ class ApplicationBase : SplitCompatApplication() {
                     networkModule,
                     appStorageModule,
                     repositoryModule,
-                    bookUseCaseModel,
                     appViewModelModule,
-                    bookViewModelModule,
                     utilityModule
                 )
             )
