+++ app/src/main/res/menu/bottom_menu.xml
@@ -15,7 +15,7 @@
         app:showAsAction="ifRoom" />
 
     <item
-        android:id="@+id/search_books"
+        android:id="@+id/search_book_graph"
         android:icon="@drawable/ic_baseline_search_24"
         android:title="@string/search"
         app:showAsAction="ifRoom" />
