+++ app/src/main/res/navigation/main_navigation.xml
@@ -6,43 +6,6 @@
     app:progressDestination="@id/progress_fragment"
     app:startDestination="@id/splashscreen">
 
-    <fragment
-        android:id="@+id/books_collections"
-        android:name="id.myone.book.presentation.book.BookFragment"
-        android:label="Home"
-        tools:layout="@layout/fragment_book">
-        <action
-            android:id="@+id/action_bookFragment_to_detailBookFragment"
-            app:destination="@id/book_details"
-            app:popUpToInclusive="true"
-            app:popUpToSaveState="false"
-            app:restoreState="false" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/book_details"
-        android:name="id.myone.book.presentation.detailBook.DetailBookFragment"
-        android:label="Detail book"
-        tools:layout="@layout/fragment_detail_book">
-        <argument
-            android:name="BOOK_ID"
-            app:argType="string" />
-        <deepLink app:uri="books-app://book-detail/{BOOK_ID}" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/search_books"
-        android:name="id.myone.book.presentation.search.SearchFragment"
-        android:label="Search book"
-        tools:layout="@layout/fragment_search">
-        <action
-            android:id="@+id/action_search_books_to_book_details"
-            app:destination="@id/book_details">
-            <argument
-                android:name="BOOK_ID"
-                app:argType="string" />
-        </action>
-    </fragment>
 
     <fragment
         android:id="@+id/splashscreen"
@@ -67,6 +30,18 @@
         android:name="id.myone.capstone_books_store.view.ProgressFragment"
         android:label="ProgressFragment"
         tools:layout="@layout/fragment_progress" />
+
+    <include-dynamic
+        android:id="@+id/books_collections"
+        app:graphPackage="id.myone.capstone_books_store.book"
+        app:graphResName="book_navigation"
+        app:moduleName="book" />
+
+    <include-dynamic
+        android:id="@+id/search_book_graph"
+        app:graphPackage="id.myone.capstone_books_store.search_book"
+        app:graphResName="search_book_graph"
+        app:moduleName="search_book" />
 
     <include-dynamic
         android:id="@+id/onboarding"
