+++ book/src/main/java/id/myone/capstone_books_store/book/domain/usecase/GetBookFavoriteStatusUseCase.kt
@@ -1,0 +1,14 @@
+/**
+ * Created by Mahmud on 16/01/23.
+ * mahmud120398@gmail.com
+ */
+
+package id.myone.capstone_books_store.book.domain.usecase
+
+import id.myone.core.domain.repository.Repository
+
+class GetBookFavoriteStatusUseCase(private val repository: Repository) {
+    suspend operator fun invoke(bookId: String): Boolean {
+        return repository.getIsInFavoriteBook(bookId)
+    }
+}
