+++ book/src/main/res/layout/fragment_book.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.book.BookFragment">
+    tools:context="id.myone.capstone_books_store.book.presentation.book.BookFragment">
 
     <LinearLayout
         android:id="@+id/linearLayout"
