+++ book/src/main/res/layout/fragment_detail_book.xml
@@ -5,7 +5,7 @@
     android:id="@+id/refresh_detail"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.detailBook.DetailBookFragment">
+    tools:context="id.myone.capstone_books_store.book.presentation.detailBook.DetailBookFragment">
 
     <include
         android:id="@+id/loading"
