--- book/src/main/res/navigation-v23/book_navigation.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/navigation_book"
-    app:startDestination="@id/books_collections">
-
-    <fragment
-        android:id="@+id/books_collections"
-        android:name="id.myone.book.presentation.book.BookFragment"
-        android:label="fragment_book"
-        tools:layout="@layout/fragment_book">
-        <action
-            android:id="@+id/action_bookFragment_to_detailBookFragment"
-            app:destination="@id/book_details"
-            app:popUpTo="@id/book_details" />
-        <deepLink
-            android:autoVerify="true"
-            app:uri="books-app://books-collection" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/book_details"
-        android:name="id.myone.book.presentation.detailBook.DetailBookFragment"
-        android:label="fragment_detail_book"
-        tools:layout="@layout/fragment_detail_book">
-        <argument
-            android:name="book_id"
-            app:argType="string" />
-
-        <deepLink app:uri="books-app://book-detail/{book_id}" />
-
-    </fragment>
-
-    <fragment
-        android:id="@+id/search_books"
-        android:name="id.myone.book.presentation.search.SearchFragment"
-        android:label="fragment_search"
-        tools:layout="@layout/fragment_search">
-        <action
-            android:id="@+id/action_search_books_to_book_details"
-            app:destination="@id/book_details">
-            <argument
-                android:name="book_id"
-                app:argType="string" />
-        </action>
-    </fragment>
-</navigation>
