+++ book/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <!-- TODO: Remove or change this placeholder text -->
+    <string name="book">Book</string>
     <string name="book_placeholder">Book Placeholder</string>
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="search_book">Search Book</string>
