+++ search_book/build.gradle
@@ -6,19 +6,17 @@ plugins {
     id 'androidx.navigation.safeargs'
 }
 
-apply from: "../share_dependencies.gradle"
+
+apply from: '../share_dependencies.gradle'
 apply from: "../share_navigation_dependencies.gradle"
 
 android {
-    namespace 'id.myone.book'
+    namespace 'id.myone.capstone_books_store.search_book'
     compileSdk 33
 
     defaultConfig {
         minSdk 21
-        targetSdk 33
-
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
@@ -28,22 +26,14 @@ android {
         }
     }
 
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-
     buildFeatures {
-        viewBinding = true
+        viewBinding=true
     }
 }
 
 dependencies {
-    implementation project(path: ":app")
     implementation project(path: ':core')
+    implementation project(path: ':app')
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 }
