+++ search_book/src/main/res/navigation/search_book_graph.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    app:startDestination="@id/search_book">
+
+    <fragment
+        android:id="@+id/search_book"
+        android:name="id.myone.capstone_books_store.search_book.presentation.SearchFragment"
+        android:label="search-book"
+        app:moduleName="search_book"
+        tools:layout="@layout/fragment_search" />
+</navigation>
