+++ settings.gradle
@@ -20,3 +20,4 @@ include ':core'
 include ':favorite'
 include ':book'
 include ':onboarding'
+include ':search_book'
